Fix attachment rendering
In Cucumber JVM attachments can be created in a few different ways. How
influences whether the attachment ends up as elements of `embeddings` or
`output`.

| Method                         | Encoding | Media Type                | Destination  |
| ------------------------------ | -------- | ------------------------- | ------------ |
| `Scenario.attach(byte[], ...)` | BASE64   | any, from method          | `embeddings` |
| `Scenario.attach(String, ...)` | IDENTITY | any, from method          | `embeddings` |
| `Scenario.log(String)`         | IDENTITY | text/x.cucumber.log+plain | `output`     |

Originally the implementation determined the destination based on the
encoding, but this is obviously not correct. An alternative would be to
select on the media type.

This does come with the disadvantage that `Scenario.attach` also takes
in a media type argument. So in theory that could be
`text/x.cucumber.log+plain` but for now we'll assume that is not the
case.

Should this assumption prove wrong we could match on both encoding and
media type, but it would be more reasonable for the user to change their
code.

Fixes: https://github.com/cucumber/cucumber-jvm/issues/3069